Rename Character to Player Add sound and music manager
