--- core/src/main/java/Assembly/Enjoyers/PlayerState.java
@@ -1,18 +1,0 @@
-package Assembly.Enjoyers;
-
-/**
- * Представляє можливі стани гравця у грі.
- * Ці стани використовуються для керування логікою анімації.
- */
-public enum PlayerState {
-    IDLE,
-    RUNNING,
-    JUMPING,
-    FALLING,
-    DASHING,
-    WALL_CLIMBING,
-    WALL_SLIDING,
-    WALL_GRABBING,
-    ATTACKING,
-    DAMAGED
-}
